backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
better-exceptions==0.3.3
python-multipart==0.0.12
python-dotenv==1.0.1
asyncpg==0.30.0
pydantic[email]==2.9.2